+++ examples/container_engine/networking.tf
@@ -1,6 +1,6 @@
 // Copyright (c) 2017, 2019, Oracle and/or its affiliates. All rights reserved.
 
-resource "oci_core_virtual_network" "test_vcn" {
+resource "oci_core_vcn" "test_vcn" {
   cidr_block     = "10.0.0.0/16"
   compartment_id = "${var.compartment_ocid}"
   display_name   = "tfVcnForClusters"
@@ -9,12 +9,12 @@ resource "oci_core_virtual_network" "test_vcn" {
 resource "oci_core_internet_gateway" "test_ig" {
   compartment_id = "${var.compartment_ocid}"
   display_name   = "tfClusterInternetGateway"
-  vcn_id         = "${oci_core_virtual_network.test_vcn.id}"
+  vcn_id         = "${oci_core_vcn.test_vcn.id}"
 }
 
 resource "oci_core_route_table" "test_route_table" {
   compartment_id = "${var.compartment_ocid}"
-  vcn_id         = "${oci_core_virtual_network.test_vcn.id}"
+  vcn_id         = "${oci_core_vcn.test_vcn.id}"
   display_name   = "tfClustersRouteTable"
 
   route_rules {
@@ -29,10 +29,10 @@ resource "oci_core_subnet" "clusterSubnet_1" {
   availability_domain = "${data.oci_identity_availability_domain.ad1.name}"
   cidr_block          = "10.0.20.0/24"
   compartment_id      = "${var.compartment_ocid}"
-  vcn_id              = "${oci_core_virtual_network.test_vcn.id}"
+  vcn_id              = "${oci_core_vcn.test_vcn.id}"
 
   # Provider code tries to maintain compatibility with old versions.
-  security_list_ids = ["${oci_core_virtual_network.test_vcn.default_security_list_id}"]
+  security_list_ids = ["${oci_core_vcn.test_vcn.default_security_list_id}"]
   display_name      = "tfSubNet1ForClusters"
   route_table_id    = "${oci_core_route_table.test_route_table.id}"
 }
@@ -42,11 +42,11 @@ resource "oci_core_subnet" "clusterSubnet_2" {
   availability_domain = "${data.oci_identity_availability_domain.ad2.name}"
   cidr_block          = "10.0.21.0/24"
   compartment_id      = "${var.compartment_ocid}"
-  vcn_id              = "${oci_core_virtual_network.test_vcn.id}"
+  vcn_id              = "${oci_core_vcn.test_vcn.id}"
   display_name        = "tfSubNet1ForClusters"
 
   # Provider code tries to maintain compatibility with old versions.
-  security_list_ids = ["${oci_core_virtual_network.test_vcn.default_security_list_id}"]
+  security_list_ids = ["${oci_core_vcn.test_vcn.default_security_list_id}"]
   route_table_id    = "${oci_core_route_table.test_route_table.id}"
 }
 
@@ -55,10 +55,10 @@ resource "oci_core_subnet" "nodePool_Subnet_1" {
   availability_domain = "${data.oci_identity_availability_domain.ad1.name}"
   cidr_block          = "10.0.22.0/24"
   compartment_id      = "${var.compartment_ocid}"
-  vcn_id              = "${oci_core_virtual_network.test_vcn.id}"
+  vcn_id              = "${oci_core_vcn.test_vcn.id}"
 
   # Provider code tries to maintain compatibility with old versions.
-  security_list_ids = ["${oci_core_virtual_network.test_vcn.default_security_list_id}"]
+  security_list_ids = ["${oci_core_vcn.test_vcn.default_security_list_id}"]
   display_name      = "tfSubNet1ForNodePool"
   route_table_id    = "${oci_core_route_table.test_route_table.id}"
 }
@@ -68,10 +68,10 @@ resource "oci_core_subnet" "nodePool_Subnet_2" {
   availability_domain = "${data.oci_identity_availability_domain.ad2.name}"
   cidr_block          = "10.0.23.0/24"
   compartment_id      = "${var.compartment_ocid}"
-  vcn_id              = "${oci_core_virtual_network.test_vcn.id}"
+  vcn_id              = "${oci_core_vcn.test_vcn.id}"
 
   # Provider code tries to maintain compatibility with old versions.
-  security_list_ids = ["${oci_core_virtual_network.test_vcn.default_security_list_id}"]
+  security_list_ids = ["${oci_core_vcn.test_vcn.default_security_list_id}"]
   display_name      = "tfSubNet2ForNodePool"
   route_table_id    = "${oci_core_route_table.test_route_table.id}"
 }