+++ examples/load_balancer/lb_private/lb_private.tf
@@ -26,7 +26,7 @@ data "oci_identity_availability_domain" "ad" {
 
 /* Network */
 
-resource "oci_core_virtual_network" "vcn1" {
+resource "oci_core_vcn" "vcn1" {
   cidr_block     = "10.1.0.0/16"
   compartment_id = "${var.compartment_ocid}"
   display_name   = "vcn1"
@@ -39,10 +39,10 @@ resource "oci_core_subnet" "subnet1" {
   display_name               = "subnet1"
   dns_label                  = "subnet1"
   compartment_id             = "${var.compartment_ocid}"
-  vcn_id                     = "${oci_core_virtual_network.vcn1.id}"
-  security_list_ids          = ["${oci_core_virtual_network.vcn1.default_security_list_id}"]
-  route_table_id             = "${oci_core_virtual_network.vcn1.default_route_table_id}"
-  dhcp_options_id            = "${oci_core_virtual_network.vcn1.default_dhcp_options_id}"
+  vcn_id                     = "${oci_core_vcn.vcn1.id}"
+  security_list_ids          = ["${oci_core_vcn.vcn1.default_security_list_id}"]
+  route_table_id             = "${oci_core_vcn.vcn1.default_route_table_id}"
+  dhcp_options_id            = "${oci_core_vcn.vcn1.default_dhcp_options_id}"
   prohibit_public_ip_on_vnic = true
 
   provisioner "local-exec" {
@@ -109,7 +109,7 @@ resource "oci_load_balancer_backend_set" "lb-bes2" {
 resource "oci_core_network_security_group" "test_network_security_group" {
   #Required
   compartment_id = "${var.compartment_ocid}"
-  vcn_id         = "${oci_core_virtual_network.vcn1.id}"
+  vcn_id         = "${oci_core_vcn.vcn1.id}"
 }
 
 output "lb_private_ip" {