+++ examples/pic/launch_by_subscription/network.tf
@@ -1,6 +1,6 @@
 // Copyright (c) 2017, 2019, Oracle and/or its affiliates. All rights reserved.
 
-resource "oci_core_virtual_network" "pic_vcn" {
+resource "oci_core_vcn" "pic_vcn" {
   cidr_block     = "10.1.0.0/16"
   compartment_id = "${var.compartment_ocid}"
   display_name   = "PICVcn"
@@ -12,22 +12,22 @@ resource "oci_core_subnet" "pic_subnet" {
   cidr_block          = "10.1.20.0/24"
   display_name        = "PICSubnet"
   dns_label           = "PICSubnet"
-  security_list_ids   = ["${oci_core_virtual_network.pic_vcn.default_security_list_id}"]
+  security_list_ids   = ["${oci_core_vcn.pic_vcn.default_security_list_id}"]
   compartment_id      = "${var.compartment_ocid}"
-  vcn_id              = "${oci_core_virtual_network.pic_vcn.id}"
+  vcn_id              = "${oci_core_vcn.pic_vcn.id}"
   route_table_id      = "${oci_core_route_table.pic_rt.id}"
-  dhcp_options_id     = "${oci_core_virtual_network.pic_vcn.default_dhcp_options_id}"
+  dhcp_options_id     = "${oci_core_vcn.pic_vcn.default_dhcp_options_id}"
 }
 
 resource "oci_core_internet_gateway" "pic_ig" {
   compartment_id = "${var.compartment_ocid}"
   display_name   = "PICIG"
-  vcn_id         = "${oci_core_virtual_network.pic_vcn.id}"
+  vcn_id         = "${oci_core_vcn.pic_vcn.id}"
 }
 
 resource "oci_core_route_table" "pic_rt" {
   compartment_id = "${var.compartment_ocid}"
-  vcn_id         = "${oci_core_virtual_network.pic_vcn.id}"
+  vcn_id         = "${oci_core_vcn.pic_vcn.id}"
   display_name   = "PICRT"
 
   route_rules {